Move util script into it's own directory and update dependent files
