+++ data/clickprofiledata.js
@@ -4,7 +4,7 @@ Utility script used to generate initial Click Profile data to populate a dataset
 
 const fs = require('fs');
 const path = require('path');
-const util = require('./util.js');
+const util = require('../util/util.js');
 
 const LOG_TAG = 'ClickProfile';
